Changed: Replaced Material flag MATF_GLOW with a glow strength multiplier (member of Material) and reworked handling of the per-material glowing properties to allow for variable strength glows. Removed per-plane glowing properties as these are now redundant. Updated game Material definitions accordingly with suitable defaults. Added: Introduced a shadow color to fakeradio. Currently shadows are colored using the inverse-amplified, color average of the surface material's primary texture. Added: Use the new fakeradio shadow color property to implement a subtle inner-edge glow to glowing surfaces. Changed: cvar rend-glow now represented in floating-point and is used as a material glow strength multiplier (default: 1).

+++ doomsday/engine/portable/include/r_data.h
@@ -239,7 +239,7 @@ extern nodepile_t* mobjNodes, *lineNodes;
 
 extern int viewwidth, viewheight;
 extern int levelFullBright;
-extern int glowingTextures;
+extern float glowingTextures;
 extern byte precacheSprites, precacheSkins;
 
 extern int numFlats;
